user auth mockup test

+++ services/player/auth/app.py
@@ -1,8 +1,13 @@
 import base64
 
 from flask import Flask
-
-from shared.auth_middleware import *
+import os
+mockup = os.getenv("MOCKUP", "0") == "1"
+gigio=None
+if mockup:
+    from auth_middleware import *
+else:
+    from shared.auth_middleware import *
 
 app = Flask(__name__)
 
@@ -21,6 +26,7 @@ def login():
         "username": sanitize(username),
         "password": password
     }
+    if mockup: return send_response(gigio("login", **data), 200)
     response = requests.post(url,  timeout=30, verify=False, json=data)
     return send_response(response.json(), response.status_code)
 
@@ -40,6 +46,7 @@ def register():
         "email": email,
         "image": base64.b64encode(image).decode('utf-8') if image else None
     }
+    if mockup: return send_response(gigio("register", **data), 200)
     response = requests.post(url,  timeout=30, verify=False, data=data)
     return send_response(response.json(), response.status_code)
 
@@ -48,6 +55,11 @@ def register():
 @token_required_void
 def logout():
     url = f"{dbm_url}/logout"
+
+    if mockup: 
+        #grab from the request header the token
+        token = request.headers["Authorization"].split(" ")[1]
+        return send_response(gigio("logout",session_token=token), 200)
     response = requests.delete(url,  timeout=30, verify=False, headers=request.headers)
     return send_response(response.json(), response.status_code)
 
@@ -60,6 +72,7 @@ def delete():
     user_id = token['user_id']
     if token['user_type'] != "PLAYER":
         user_id = request.args.get('user_id')
+    if mockup: return send_response(gigio("delete", user_id=user_id),200)
     url = f"{dbm_url}/delete/PLAYER"
     response = requests.delete(url+f'?user_id={user_id}', timeout=30, verify=False, headers=generate_session_token_system())
     return send_response(response.json(), response.status_code)
